Extracted stop loss state initialization, fixed issue with initial state (#972)
* extracted stop loss state initialization

* fixes after dev merge

* removed console log
