+++ components/vault/GeneralManageLayout.tsx
@@ -1,26 +1,14 @@
 import { useTranslation } from 'next-i18next'
-import React, { useEffect } from 'react'
+import React from 'react'
 import { Grid } from 'theme-ui'
 
 import { ALLOWED_MULTIPLY_TOKENS } from '../../blockchain/tokensMetadata'
-import {
-  getInitialVaultCollRatio,
-  getStartingSlRatio,
-} from '../../features/automation/protection/common/helpers'
-import { extractStopLossData } from '../../features/automation/protection/common/StopLossTriggerDataExtractor'
-import { ADD_FORM_CHANGE } from '../../features/automation/protection/common/UITypes/AddFormChange'
-import {
-  PROTECTION_MODE_CHANGE_SUBJECT,
-  ProtectionModeChange,
-} from '../../features/automation/protection/common/UITypes/ProtectionFormModeChange'
-import { REMOVE_FORM_CHANGE } from '../../features/automation/protection/common/UITypes/RemoveFormChange'
 import { TriggersData } from '../../features/automation/protection/triggers/AutomationTriggersData'
+import { useStopLossStateInitializator } from '../../features/automation/protection/useStopLossStateInitializator'
 import { VaultBannersView } from '../../features/banners/VaultsBannersView'
 import { GeneralManageVaultState } from '../../features/generalManageVault/generalManageVault'
 import { GeneralManageVaultViewAutomation } from '../../features/generalManageVault/GeneralManageVaultView'
-import { useUIChanges } from '../../helpers/uiChangesHook'
 import { useFeatureToggle } from '../../helpers/useFeatureToggle'
-import { useAppContext } from '../AppContextProvider'
 import { VaultTabSwitch, VaultViewMode } from '../VaultTabSwitch'
 import { DefaultVaultHeaderControl } from './DefaultVaultHeaderControl'
 import { HistoryControl } from './HistoryControl'
@@ -38,55 +26,10 @@ export function GeneralManageLayout({
   autoTriggersData,
 }: GeneralManageLayoutProps) {
   const { t } = useTranslation()
-  const { uiChanges } = useAppContext()
   const { ilkData, vault, account, priceInfo } = generalManageVault.state
   const showProtectionTab = ALLOWED_MULTIPLY_TOKENS.includes(vault.token)
-  const { stopLossLevel, isStopLossEnabled, isToCollateral } = extractStopLossData(autoTriggersData)
-  const [currentForm] = useUIChanges<ProtectionModeChange>(PROTECTION_MODE_CHANGE_SUBJECT)
   const automationBasicBuyAndSellEnabled = useFeatureToggle('AutomationBasicBuyAndSell')
-
-  const initialVaultCollRatio = getInitialVaultCollRatio({
-    liquidationRatio: generalManageVault.state.ilkData.liquidationRatio,
-    collateralizationRatio: generalManageVault.state.vault.collateralizationRatio,
-  })
-
-  const startingSlRatio = getStartingSlRatio({
-    stopLossLevel,
-    isStopLossEnabled,
-    initialVaultCollRatio,
-  })
-
-  useEffect(() => {
-    uiChanges.publish(ADD_FORM_CHANGE, {
-      type: 'close-type',
-      toCollateral: isToCollateral,
-    })
-    uiChanges.publish(ADD_FORM_CHANGE, {
-      type: 'stop-loss',
-      stopLoss: startingSlRatio.multipliedBy(100),
-    })
-    uiChanges.publish(ADD_FORM_CHANGE, {
-      type: 'tx-details',
-      txDetails: {},
-    })
-    uiChanges.publish(REMOVE_FORM_CHANGE, {
-      type: 'tx-details',
-      txDetails: {},
-    })
-  }, [currentForm])
-
-  useEffect(() => {
-    uiChanges.publish(ADD_FORM_CHANGE, {
-      type: 'close-type',
-      toCollateral: isToCollateral,
-    })
-    uiChanges.publish(ADD_FORM_CHANGE, {
-      type: 'stop-loss',
-      stopLoss: startingSlRatio.multipliedBy(100),
-    })
-  }, [isStopLossEnabled])
-
-  const protectionEnabled = !!generalManageVault.state.stopLossData?.isStopLossEnabled
+  const isStopLossEnabled = useStopLossStateInitializator(ilkData, vault, autoTriggersData)
 
   return (
     <Grid gap={0} sx={{ width: '100%' }}>
@@ -116,7 +59,7 @@ export function GeneralManageLayout({
         protectionControl={<ProtectionControl vault={vault} ilkData={ilkData} account={account} />}
         vaultInfo={<VaultInformationControl generalManageVault={generalManageVault} />}
         showProtectionTab={showProtectionTab}
-        protectionEnabled={protectionEnabled}
+        protectionEnabled={isStopLossEnabled}
       />
     </Grid>
   )
